Zhuohengw

+++ Bruha/LoadScreenViewController.swift
@@ -19,12 +19,15 @@ class LoadScreenViewController: UIViewController {
     var generalRetrieved = 0
     var loggedinRetrieved = 0
     
+    var userLog: Int = 0
     
     override func viewDidLoad() {
         super.viewDidLoad()
         progressView.tintColor = UIColor.purpleColor()
         progressView.progressViewStyle = UIProgressViewStyle.Bar
         progressView.backgroundColor = UIColor.orangeColor()
+        
+        userLog = (FetchData(context: managedObjectContext).fetchUserInfo()?.count)!
         
 //        DeleteData(context: self.managedObjectContext).deleteAll()
 //        LoadScreenService(context: self.managedObjectContext).retrieveAll()
@@ -33,29 +36,32 @@ class LoadScreenViewController: UIViewController {
     override func viewWillAppear(animated: Bool) {
         super.viewWillAppear(animated)
         
-        DeleteData(context: self.managedObjectContext).deleteAll()
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Event Categories", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Events", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Events", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Events", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Venues", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Venues", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Venues", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Organizations", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Organizations", object: nil)
-        
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Organizations", object: nil)
-        
-        LoadScreenService(context: self.managedObjectContext).retrieveAll()
+        if hasInternet() == "true" {
+            
+            DeleteData(context: self.managedObjectContext).deleteAll()
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Event Categories", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Events", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Events", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Events", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Venues", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Venues", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Venues", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueSplash", name:"Organizations", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"User Organizations", object: nil)
+            
+            NSNotificationCenter.defaultCenter().addObserver(self, selector: "segueDashBoard", name:"Addicted Organizations", object: nil)
+            
+            LoadScreenService(context: self.managedObjectContext).retrieveAll()
+        }
     }
     
     override func viewDidAppear(animated: Bool) {
@@ -74,7 +80,7 @@ class LoadScreenViewController: UIViewController {
             
             let okAction = UIAlertAction(title: "OK", style: .Default) { (_) -> Void in
                 
-                if (FetchData(context: self.managedObjectContext).fetchUserInfo()?.count != 0) {
+                if (self.userLog != 0) {
                     
                     GlobalVariables.loggedIn = true
                 }
@@ -129,9 +135,10 @@ class LoadScreenViewController: UIViewController {
                 
                 self.getDataFromUrl(checkedUrl) { data in
                     dispatch_async(dispatch_get_main_queue()) {
-                        if let downloadImg = UIImage(data: data!) {
-                            GlobalVariables.ImageArray.append((event.eventID, downloadImg))
+                        if let downloadImg = data {
                             //print("event")
+                            let img: Image = Image(id: event.eventID, image: downloadImg)
+                            SaveData(context: self.managedObjectContext).savePosterImages(img)
                         }
                         
                     }
@@ -143,8 +150,9 @@ class LoadScreenViewController: UIViewController {
                 
                 self.getDataFromUrl(checkedUrl) { data in
                     dispatch_async(dispatch_get_main_queue()) {
-                        if let downloadImg = UIImage(data: data!) {
-                            GlobalVariables.ImageArray.append((venue.venueID, downloadImg))
+                        if let downloadImg = data {
+                            let img: Image = Image(id: venue.venueID, image: downloadImg)
+                            SaveData(context: self.managedObjectContext).savePosterImages(img)
                         }
                     }
                 }
@@ -155,8 +163,9 @@ class LoadScreenViewController: UIViewController {
                 
                 self.getDataFromUrl(checkedUrl) { data in
                     dispatch_async(dispatch_get_main_queue()) {
-                        if let downloadImg = UIImage(data: data!) {
-                            GlobalVariables.ImageArray.append((organization.organizationID, downloadImg))
+                        if let downloadImg = data {
+                            let img: Image = Image(id: organization.organizationID, image: downloadImg)
+                            SaveData(context: self.managedObjectContext).savePosterImages(img)
                         }
                     }
                 }
@@ -192,7 +201,7 @@ class LoadScreenViewController: UIViewController {
         
         generalRetrieved = generalRetrieved + 1
         
-        if generalRetrieved == 4 && FetchData(context: managedObjectContext).fetchUserInfo()?.count == 0 {
+        if generalRetrieved == 4 && userLog == 0 {
             
             print("should go to splash")
             NSNotificationCenter.defaultCenter().postNotificationName("complete", object: nil)
@@ -208,7 +217,7 @@ class LoadScreenViewController: UIViewController {
         
         loggedinRetrieved = loggedinRetrieved + 1
         
-        if loggedinRetrieved == 6 && FetchData(context: managedObjectContext).fetchUserInfo()?.count != 0 && generalRetrieved == 4 {
+        if loggedinRetrieved == 6 && userLog != 0 && generalRetrieved == 4 {
             
             print("should go to dashboard")
             NSNotificationCenter.defaultCenter().postNotificationName("complete", object: nil)
@@ -219,7 +228,7 @@ class LoadScreenViewController: UIViewController {
         
         finishesLoadingProgress()
         
-        if(FetchData(context: managedObjectContext).fetchUserInfo()?.count != 0) {
+        if(userLog != 0) {
             
             delay(1.5){
                 print("DashBoard segue")
@@ -260,7 +269,7 @@ class LoadScreenViewController: UIViewController {
     
     func timeOutCheck() {
         
-        if !(loggedinRetrieved == 6 && generalRetrieved == 4 || generalRetrieved == 4 && FetchData(context: managedObjectContext).fetchUserInfo()?.count == 0) {
+        if !(loggedinRetrieved == 6 && generalRetrieved == 4 || generalRetrieved == 4 && userLog == 0) {
             
             let alert = UIAlertView(title: "Error while loading, some functionality may not work, please restart the app to get full functionality", message: nil, delegate: nil, cancelButtonTitle: nil)
             alert.show()
